third_party/WebKit/LayoutTests/fast/canvas/offscreencanvas-text/text/2d.text.font.parse.size.percentage.default.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.font.parse.size.percentage.default</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/external/wpt/common/canvas-tests.js"></script>

<h1>2d.text.font.parse.size.percentage.default</h1>
<p class="desc"></p>


<script>
var t = async_test("");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var canvas2 = document.createElement('canvas');
var ctx2 = canvas2.getContext('2d');
ctx2.font = '1000% serif';
_assertSame(ctx2.font, '100px serif', "ctx2.font", "'100px serif'");

t.done();

});
</script>
